backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pydantic-settings
